# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NaikPete' API",
        "description": "API documentation for My Application",
        "termsOfService": "https://www.example.com/terms",
        "contact": {
            "email": "contact@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/get-trayek-with-halte/{trayekId}": {
            "get": {
                "tags": [
                    "Pemesanan"
                ],
                "summary": "Get trayek with halte details",
                "operationId": "7e2de877aab564150fe8583522ff43e7",
                "parameters": [
                    {
                        "name": "trayekId",
                        "in": "path",
                        "description": "Trayek ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Trayek with haltes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "kode_trayek": {
                                            "type": "string"
                                        },
                                        "urutan_halte": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Halte"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Trayek not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Trayek tidak ditemukan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/beli-tiket": {
            "post": {
                "tags": [
                    "Pemesanan"
                ],
                "summary": "Buy a ticket",
                "operationId": "4ea03dd488d4dfd51cf6d5ded76b4f91",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "trayek_id": {
                                        "type": "integer"
                                    },
                                    "jumlah_tiket": {
                                        "type": "integer"
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "enum": [
                                            "cash",
                                            "cashless"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ticket purchased successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tiket berhasil dibeli"
                                        },
                                        "ticket": {
                                            "$ref": "#/components/schemas/Ticket"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Invalid ticket or expired",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Tiket tidak valid atau sudah kadaluarsa"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/cek-status-tiket": {
            "post": {
                "tags": [
                    "Pemesanan"
                ],
                "summary": "Check ticket status",
                "operationId": "6e7b07242b7403bfcdfd12ba4db956ae",
                "responses": {
                    "200": {
                        "description": "Ticket is valid and can be used",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tiket valid dan dapat digunakan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Ticket is invalid or expired",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Tiket tidak valid atau sudah kadaluarsa"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/pesan-pete": {
            "post": {
                "tags": [
                    "Pemesanan"
                ],
                "summary": "Pesan angkot (angkutan kota)",
                "operationId": "f95d5de365c349c1d7c066adc9bea51b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "trayek_id": {
                                        "description": "ID trayek yang ingin dipesan",
                                        "type": "integer"
                                    },
                                    "jumlah_tiket": {
                                        "description": "Jumlah tiket yang ingin dipesan",
                                        "type": "integer"
                                    },
                                    "payment_method": {
                                        "description": "Metode pembayaran",
                                        "type": "string",
                                        "enum": [
                                            "cash",
                                            "cashless"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tiket berhasil dipesan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tiket berhasil dipesan"
                                        },
                                        "transaksi": {
                                            "$ref": "#/components/schemas/Transaction"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Hanya customer yang dapat memesan tiket atau tiket belum dibeli",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Hanya customer yang dapat memesan tiket"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Trayek atau kendaraan tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tidak ada kendaraan yang tersedia untuk trayek ini"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Input tidak valid atau ada kesalahan dalam pemesanan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tiket belum dibeli"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Driver": {
                "required": [
                    "id_driver",
                    "nama",
                    "email",
                    "password",
                    "alamat",
                    "tgl_lahir",
                    "no_hp",
                    "gender",
                    "id_kendaraan"
                ],
                "properties": {
                    "id_driver": {
                        "description": "ID of the driver",
                        "type": "integer"
                    },
                    "nama": {
                        "description": "Name of the driver",
                        "type": "string"
                    },
                    "email": {
                        "description": "Email of the driver",
                        "type": "string"
                    },
                    "password": {
                        "description": "Password of the driver",
                        "type": "string"
                    },
                    "alamat": {
                        "description": "Address of the driver",
                        "type": "string"
                    },
                    "tgl_lahir": {
                        "description": "Date of birth of the driver",
                        "type": "string",
                        "format": "date"
                    },
                    "no_hp": {
                        "description": "Phone number of the driver",
                        "type": "string"
                    },
                    "gender": {
                        "description": "Gender of the driver",
                        "type": "string"
                    },
                    "id_kendaraan": {
                        "description": "ID of the vehicle assigned to the driver",
                        "type": "integer"
                    },
                    "kendaraan": {
                        "$ref": "#/components/schemas/Kendaraan"
                    }
                },
                "type": "object"
            },
            "Halte": {
                "required": [
                    "id",
                    "nama_halte",
                    "latitude",
                    "longitude"
                ],
                "properties": {
                    "id": {
                        "description": "ID of the halte",
                        "type": "integer"
                    },
                    "nama_halte": {
                        "description": "Name of the halte",
                        "type": "string"
                    },
                    "latitude": {
                        "description": "Latitude of the halte",
                        "type": "number",
                        "format": "float"
                    },
                    "longitude": {
                        "description": "Longitude of the halte",
                        "type": "number",
                        "format": "float"
                    }
                },
                "type": "object"
            },
...
            "Ticket": {
                "required": [
                    "id",
                    "id_user",
                    "tgl_pembelian",
                    "expired_at",
                    "status"
                ],
                "properties": {
                    "id": {
                        "description": "ID of the ticket",
                        "type": "integer"
                    },
                    "id_user": {
                        "description": "ID of the user who purchased the ticket",
                        "type": "integer"
                    },
                    "tgl_pembelian": {
                        "description": "Ticket purchase date",
                        "type": "string",
                        "format": "date-time"
                    },
                    "expired_at": {
                        "description": "Ticket expiration date",
                        "type": "string",
                        "format": "date-time"
                    },
                    "status": {
                        "description": "Ticket status (active/expired)",
                        "type": "string"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
            "Transaction": {
                "required": [
                    "id",
                    "id_user",
                    "id_driver",
                    "id_trayek",
                    "jenis_tiket",
                    "tgl_transaksi",
                    "jumlah_tiket",
                    "payment_method",
                    "payment_amount",
                    "payment_status"
                ],
                "properties": {
                    "id": {
                        "description": "ID of the transaction",
                        "type": "integer"
                    },
                    "id_user": {
                        "description": "ID of the user who made the transaction",
                        "type": "integer"
                    },
                    "id_driver": {
                        "description": "ID of the driver involved in the transaction",
                        "type": "integer"
                    },
                    "id_trayek": {
                        "description": "ID of the route (trayek) associated with the transaction",
                        "type": "integer"
                    },
                    "jenis_tiket": {
                        "description": "Type of ticket (e.g., regular, VIP)",
                        "type": "string"
                    },
                    "tgl_transaksi": {
                        "description": "Transaction date and time",
                        "type": "string",
                        "format": "date-time"
                    },
                    "jumlah_tiket": {
                        "description": "Number of tickets purchased",
                        "type": "integer"
                    },
                    "payment_method": {
                        "description": "Payment method used (e.g., cash, credit card)",
                        "type": "string"
                    },
                    "payment_amount": {
                        "description": "Amount paid for the transaction",
                        "type": "number",
                        "format": "float"
                    },
                    "payment_status": {
                        "description": "Status of the payment (e.g., pending, completed)",
                        "type": "string"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "driver": {
                        "$ref": "#/components/schemas/Driver"
                    },
                    "trayek": {
                        "$ref": "#/components/schemas/Trayek"
                    }
                },
                "type": "object"
            },
            "Trayek": {
                "required": [
                    "id",
                    "kode_trayek",
                    "urutan_halte",
                    "id_schedule"
                ],
                "properties": {
                    "id": {
                        "description": "ID of the trayek",
                        "type": "integer"
                    },
                    "kode_trayek": {
                        "description": "Trayek code",
                        "type": "string"
                    },
                    "urutan_halte": {
                        "description": "Order of halts",
                        "type": "integer"
                    },
                    "id_schedule": {
                        "description": "Associated schedule ID",
                        "type": "integer"
                    },
                    "schedule": {
                        "$ref": "#/components/schemas/Schedule"
                    },
                    "haltes": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Halte"
                        }
                    },
                    "kendaraan": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Kendaraan"
                        }
                    },
                    "transaksi": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Transaction"
                        }
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nama": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "example": "john.doe@example.com"
                    },
                    "no_hp": {
                        "type": "string",
                        "example": "081234567890"
                    },
                    "alamat": {
                        "type": "string",
                        "example": "Jl. Merdeka No. 123"
                    },
                    "gender": {
                        "type": "string",
                        "enum": [
                            "laki-laki",
                            "perempuan"
                        ],
                        "example": "laki-laki"
                    },
                    "tgl_lahir": {
                        "type": "string",
                        "format": "date",
                        "example": "1990-01-01"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "customer",
                            "driver"
                        ],
                        "example": "customer"
                    },
                    "foto_profil": {
                        "description": "Optional profile photo",
                        "type": "string",
                        "format": "binary"
                    }
                },
                "type": "object"
            }
        }
    },
...
}